python-nlp/requirements.txt

spacy==3.7.2
transformers==4.35.0
torch==2.2.0
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0